jResp: Major infrastructural changes in project and refactoring.
This refactoring removes the sim/ folder and moves the appropiate files
to core/.

All of the examples found through the project have been gathered into
one place, and Gradle build automation scripts have been prepared.

Currently, some of the examples will fail due to a bit of lack of
scripting or missing files. This will be fixed soon.

Also, all meta-data related to Eclipse and commited binaries (except
pastry) has been extract and removed.

In order to build the core/ folder, simply do:

    gradle build

Or if one is about to modify the library in Eclipse, do:

    gradle eclipse

Then import the project as usual and start developing.

Alot of the namespace collision or package naming has also been
resolved, but this is not over yet. For now this will give us a somewhat
more stable infrastructure to work on.